requirements.txt
aioredis==1.3.1
asgiref==3.4.1
async-timeout==4.0.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
cffi==1.15.0
channels==3.0.4
channels-redis==3.3.1
constantly==15.1.0
cryptography==35.0.0
daphne==3.0.2
Django==3.2.9
django-phonenumber-field==6.0.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
hiredis==2.0.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
iniconfig==1.1.1
msgpack==1.0.2
packaging==21.2
phonenumbers==8.12.37
Pillow==8.4.0
pkg_resources==0.0.0
pluggy==1.0.0
psycopg2-binary==2.9.2
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.3.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pytest==6.2.5
pytest-asyncio==0.16.0
pytest-django==4.4.0
pytz==2021.3
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
Twisted==21.7.0
txaio==21.2.1
typing-extensions==3.10.0.2
zope.interface==5.4.0
